# openapi.yml
openapi: 3.0.0
info:
  title: Entries API
  version: 1.0.0

paths:
  /entries:
    get:
      summary: Get latest entries
      operationId: getLatestEntries
      parameters:
        - name: last_last_edited_at
          in: query
          description: Filter entries by the last edited date
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: A list of latest entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetLatestEntriesRow'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new entry
      operationId: createEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntryRequest'
      responses:
        '201':
          description: Entry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEntryResponse'
        '409':
          description: Entry with the same title already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /entries/{path}:
    get:
      summary: Get entry by dynamic path
      operationId: getEntryByDynamicPath
      parameters:
        - name: path
          in: path
          required: true
          description: The path of the entry
          schema:
            type: string
      responses:
        '200':
          description: An entry object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLatestEntriesRow'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /entries/{path}/links:
    get:
      summary: Get linked entry paths
      operationId: getLinkedEntryPaths
      parameters:
        - name: path
          in: path
          required: true
          description: The source entry path
          schema:
            type: string
      responses:
        '200':
          description: A map of linked entry titles to their paths
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedEntriesResponse'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /entries/{path}/body:
    put:
      summary: Update entry body
      operationId: updateEntryBody
      parameters:
        - name: path
          in: path
          required: true
          description: The path of the entry to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEntryBodyRequest'
      responses:
        '200':
          description: Entry content updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /entries/{path}/title:
    put:
      summary: Update entry title
      operationId: updateEntryTitle
      parameters:
        - name: path
          in: path
          required: true
          description: The path of the entry to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEntryTitleRequest'
      responses:
        '200':
          description: Entry title updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Entry with the same title already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /entries/titles:
    get:
      summary: Get all entry titles
      operationId: getAllEntryTitles
      responses:
        '200':
          description: A list of all entry titles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryTitlesResponse'
        default:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    GetLatestEntriesRow:
      type: object
      properties:
        Path:
          type: string
        Title:
          type: string
        Body:
          type: string
        Visibility:
          type: string
        Format:
          type: string
        PublishedAt:
          type: string
          format: date-time
          nullable: true
        LastEditedAt:
          type: string
          format: date-time
          nullable: true
        CreatedAt:
          type: string
          format: date-time
          nullable: true
        UpdatedAt:
          type: string
          format: date-time
          nullable: true
        ImageUrl:
          type: string
          nullable: true

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string

    LinkedEntriesResponse:
      type: object
      additionalProperties:
        type: string
        nullable: true
      description: Object where keys are lowercase destination entry titles and values are their paths (null if entry doesn't exist)

    UpdateEntryBodyRequest:
      type: object
      required:
        - body
      properties:
        body:
          type: string
          description: The new content of the entry

    UpdateEntryTitleRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The new title for the entry

    EntryTitlesResponse:
      type: array
      items:
        type: string
      description: Array of all entry titles

    CreateEntryRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The title of the new entry

    CreateEntryResponse:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          description: The path of the created entry

    EmptyResponse:
      type: object
      properties: {}